Refactor WatchConfig to panic inline instead of os.Exit() in a goroutine on init fail
Fixes #2095

I Chose to also add panics to the errors for the getConfigFile and watcher.Add as either of those failing was also returning, tho not exiting. This increases the consistency of this area of teh code.
I am unsure if it would be better to just remove the panics and rely on the user not getting updates as expected.
Unless there is some manner for listening to errors that occur in viper (an error channel etc) I think this is the best bet.